seperating the selection methods into different files and adding a basic random selection method
